[refactor/copy_texts] Changing the copy texts in different parts of the app so they are more self-explanatory.
