+++ example1/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
-    <string name="app_name">example1</string>
+    <string name="example_1_long_description">We are using the GetNewsInteractor to fetch the recently viewed news and the paginated news altogether. Notice that when we paginate through the list we don\'t fetch the recently viewed news again. However, if we do a Pull-To-Refresh then we do fetch them again.</string>
+    <string name="example_2_long_description">Here we\'re using the GetProfileInteractor to fetch the personal information and business skills. Notice that if something goes wrong while either calling the /profile or the /business, we automatically hide that section.</string>
+    <string name="example_2_toolbar_title">RxJava Example #2</string>
+    <string name="example_1_toolbar_title">RxJava Example #1</string>
 </resources>
