+++ example2/src/main/res/layout/activity_example_cache_1_detail.xml
@@ -20,7 +20,7 @@
 
             <androidx.appcompat.widget.Toolbar
                 android:id="@+id/activityExampleCache1DetailToolbar"
-                app:title="Example Cache Recipes"
+                app:title="@string/example_2_cache_detail_toolbar_title"
                 app:layout_collapseMode="pin"
                 android:theme="@style/Examples.AppBarOverlay"
                 android:layout_width="match_parent"
@@ -38,7 +38,7 @@
                     android:layout_marginStart="8dp"
                     android:layout_marginEnd="8dp"
                     android:layout_marginBottom="8dp"
-                    android:text="Here we are able to fetch a single Recipe by id from cache. We are also able to bookmark the recipe by tapping on the bookmark icon. Since we are using Flowables within the Interactor, as soon as we mark the Recipe as bookmarked, the Recipe list on the previous activity it is also updated."
+                    android:text="@string/example_2_cache_detail_long_description"
                     android:padding="4dp"
                     style="@style/Examples.TextAppearance.Description.Inverse"
                     android:layout_width="match_parent"
