+++ example3/src/main/res/layout/example_imgur_activity.xml
@@ -18,7 +18,7 @@
 
             <androidx.appcompat.widget.Toolbar
                 android:id="@+id/activityExampleImgurToolbar"
-                app:title="Example Cache Recipes"
+                app:title="@string/example_3_toolbar_title"
                 android:theme="@style/Examples.AppBarOverlay"
                 android:layout_width="match_parent"
                 android:layout_height="?attr/actionBarSize"/>
